WIP
* experiment - testing app that renders camera capture to a window.
* hello - testing XR app that shows a cube
* src/api_layer.rs - WIP passthrough layer

TODO:

we are doing undistort and yuv conversion on a separate camera thread.
in api layer, this is done with the VkQueue the application provides.
OpenXR dictates that this VkQueue can only be used in
xrAcquire/ReleaseSwapchainImage and xrBegin/EndFrame. Using it in a
separate thread is not OK.

Needs a fork of openxrs to build, not pushed anywhere or merged upstream
yet.